drf/requirements.txt
django==3.1.6
djangorestframework==3.12.2
drf-yasg==1.20.0
psycopg2_binary==2.8.6
pyyaml==5.4.1
